requirement.txt
Django==5.0.6
djangorestframework==3.15.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
setuptools==59.6.0
djangorestframework-simplejwt==5.3.1
djangorestframework
asgiref==3.8.1
sqlparse==0.4.4

